pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlops-incident-env"
version = "1.0.0"
description = "MLOps Incident Response Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]

[tool.openenv]
name = "mlops-incident-env"
version = "1.0.0"
description = "MLOps Incident Response RL Environment"
